Changed the `source` field to point at the `podspec/#.#.#` tag instead of just `#.#.#`
In the past and into the present, the `#.#.#` tags represent the source code that builds in Xcode and with Swift Package Manager, and the `podspec/#.#.#` tags represent changes made only to the podspec and other CocoaPods-related stuff to appease the every-mercurial CocoaPods spec and other requirements.  While the `#.#.#` will build just fine, the `podspec/#.#.#` represents what was actually sent to the CocoaPods trunk repo, and will lint properly with the current CocoaPods version (at date of commit).

Since uploading a pod to CocoaPods has long been a tiresome practice of tweaking variables to appease the general lack of backwards-compatibility, and these changes have nothing to do with the actual source of the project, I started doing the separate `podspec/#.#.#` versions a while back and it only seems fitting that since CocoaPods needs to feel special all the time, it can have a special non-standard repo version.