Added `rotate(by:)`/`unrotate(by:)` helpers to `Float3`
Since `FloatQuaternion` has methods to `rotate` & `unrotate` a `Float3` vector, it makes sense to be able to tell it `Float3` to `rotate`/`unrotate` itself `by:` a `FloatQuaternion`.

Decided to not add the `*` operator— it makes more sense have only the traditional `quaternion * vector` form.  I may change my mind on this later if I find there's mathematical soundness to symmetrical applying of a quaternion to a vector, but I'm pretty sure this is just a game engine shorthand without any mathematical basis.

* Added a new “Geometric Math Operations” `Float3` extension with  `rotate`(`d`)`(by:FloatQuaterion)` & `unrotate`(`d`)`(by:FloatQuaterion)` convenience methods.  
	▸ These just call `FloatQuaterion`'s `rotate(…)`/`unrotate(…)` methods.